+++ spec/fixtures/repos/assets_in_hook/deploy/before_compile_assets.rb
@@ -1,2 +1,2 @@
-run 'touch custom_compiled && mkdir -p public/assets'
+run 'touch custom_compiled && mkdir -p public/assets && touch public/assets/custom_compiled_asset'
 $stderr.puts Dir.pwd