+++ spec/rails31_deploy_spec.rb
@@ -2,18 +2,25 @@ require 'spec_helper'
 
 describe "Deploying a Rails 3.1 application" do
   context "with default production settings" do
-    it "precompiles assets, then reuses them on the next deploy if nothing has changed" do
-      deploy_test_application('assets_enabled')
+    it "precompiles assets when asset compilation is detected" do
+      deploy_test_application('assets_detected')
       deploy_dir.join('current', 'precompiled').should exist
       deploy_dir.join('current', 'public', 'assets').should exist
       deploy_dir.join('current', 'public', 'assets', 'compiled_asset').should exist
       read_output.should include("Precompiling assets. ('app/assets' exists, 'public/assets' not found, not disabled in config.)")
+    end
+
+    it "precompiles assets, then reuses them on the next deploy if nothing has changed" do
+      deploy_test_application('assets_enabled_in_ey_yml')
+      deploy_dir.join('current', 'precompiled').should exist
+      deploy_dir.join('current', 'public', 'assets').should exist
+      deploy_dir.join('current', 'public', 'assets', 'compiled_asset').should exist
 
       redeploy_test_application
       deploy_dir.join('current', 'precompiled').should_not exist # doesn't run the task
       deploy_dir.join('current', 'public', 'assets').should exist # but the assets are there
       deploy_dir.join('current', 'public', 'assets', 'compiled_asset').should exist
-      read_output.should =~ %r#Reusing existing assets\. \('app/assets' unchanged from \w{7}..\w{7}\)#
+      read_output.should =~ %r#Reusing existing assets\. \(configured asset_dependencies unchanged from \w{7}..\w{7}\)#
 
       redeploy_test_application('config' => {'precompile_unchanged_assets' => 'true'})
       deploy_dir.join('current', 'precompiled').should exist # doesn't run the task
@@ -23,18 +30,18 @@ describe "Deploying a Rails 3.1 application" do
     end
 
     it "precompile assets again when redeploying a ref with changes" do
-      deploy_test_application('assets_enabled')
+      deploy_test_application('assets_enabled_in_ey_yml')
       deploy_dir.join('current', 'precompiled').should exist
       deploy_dir.join('current', 'public', 'assets').should exist
       deploy_dir.join('current', 'public', 'assets', 'compiled_asset').should exist
-      read_output.should include("Precompiling assets. ('app/assets' exists, 'public/assets' not found, not disabled in config.)")
+      read_output.should include("Precompiling assets. (precompile_assets: true)")
 
       # changing the ref stands in for actually having assets change (see Strategies::IntegrationSpec#same?)
       redeploy_test_application('branch' => 'somenewref')
       deploy_dir.join('current', 'precompiled').should exist # it does runs the task
       deploy_dir.join('current', 'public', 'assets').should exist
       deploy_dir.join('current', 'public', 'assets', 'compiled_asset').should exist
-      read_output.should_not include("Reusing existing assets. (assets appear to be unchanged)")
+      read_output.should_not =~ %r#Reusing existing assets#
     end
   end
 
@@ -103,6 +110,7 @@ describe "Deploying a Rails 3.1 application" do
       deploy_dir.join('current', 'precompiled').should_not exist
       deploy_dir.join('current', 'public', 'assets').should be_directory
       deploy_dir.join('current', 'public', 'assets').should_not be_symlink
+      deploy_dir.join('current', 'public', 'assets', 'custom_compiled_asset').should exist
       read_output.should include("Skipping asset precompilation. ('public/assets' directory already exists.)")
     end
   end